Moved the internally correct but unused caching classes to their own package.
