+++ src/opendap/caching/BesCatalogCache.java
@@ -1,5 +1,7 @@
-package opendap.bes;
+package opendap.caching;
 
+import opendap.bes.BESError;
+import opendap.bes.BadConfigurationException;
 import opendap.bes.dap2Responders.BesApi;
 import opendap.ppt.PPTException;
 import org.jdom.Document;
