+++ src/opendap/caching/BesResponseCache.java
@@ -1,4 +1,4 @@
-package opendap.bes;
+package opendap.caching;
 
 import org.slf4j.Logger;
